+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.3.10
+
+### Patch Changes
+
+-   Updated dependencies [f8e16120]
+    -   @sap-ux/project-access@1.19.10
+
+## 0.3.9
+
+### Patch Changes
+
+-   Updated dependencies [ee76e47f]
+    -   @sap-ux/project-access@1.19.9
+
+## 0.3.8
+
+### Patch Changes
+
+-   Updated dependencies [f8b24a3a]
+    -   @sap-ux/inquirer-common@0.2.6
+
+## 0.3.7
+
+### Patch Changes
+
+-   Updated dependencies [148d011a]
+    -   @sap-ux/ui5-info@0.4.0
+    -   @sap-ux/inquirer-common@0.2.5
+
+## 0.3.6
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.8
+
 ## 0.3.5
 
 ### Patch Changes